+++ changelog.md
@@ -1,5 +1,12 @@
 [![Downloads](https://img.shields.io/github/downloads/probablyraging/steam-game-idler/1.5.28/total?style=for-the-badge&logo=github&color=137eb5)](https://github.com/probablyraging/steam-game-idler/releases/download/1.5.28/Steam.Game.Idler_1.5.28_x64_en-US.msi)
 
 ## Changelog
-- Replaced the individual `Update` buttons with a single `Save changes` button on the `Achievement > Statistics` tab
-  - (#36) by @Jesewe 
+- Added an `auto idle games` feature.
+  - This works similar to the `card farming` and `achievement unlocker` lists
+  - Add a max of 32 games to the `auto idle` list by clicking the 3 vertical dots on a game's card
+  - View games in the `auto idle` list by choosing it from the drop down `filter` menu
+  - When SGI launches, all games in this list will be automatically idled
+    - (#35, #48) by @Nevenit, @TiimmyTuurner
+- Added a `steamParental/steamMachineAuth` input field to `settings > card farming`
+  - **This field is optional** and is **only required if** a `steamParental` and/or `steamMachineAuth` cookie is present when following [these steps](https://github.com/probablyraging/steam-game-idler/wiki/Settings#steam-credentials)
+    - (#49) by @FireLostBoy-Tech